+++ src/app/modules/slorm-reaper/components/reaper-list/reaper-list.component.ts
@@ -8,6 +8,8 @@ import { toBlob } from 'html-to-image';
 import { combineLatest, takeUntil } from 'rxjs';
 import {
     HeroClass,
+    MAX_EFFECT_AFFINITY_BASE,
+    MAX_REAPER_AFFINITY_BASE,
     MAX_REAPER_AFFINITY_BONUS,
     Reaper,
     SlormancerDataService,
@@ -80,7 +82,18 @@ export class ReaperListComponent extends AbstractUnsubscribeComponent implements
 
     private buildReaperList(heroClass: HeroClass, primordial: boolean, maxLevelAndAffinity: boolean) {
         this.allReapers = this.slormancerDataService.getGameDataAvailableReaper()
-            .map(reaperData => this.slormancerReaperService.getReaper(reaperData, heroClass, primordial, maxLevelAndAffinity ? valueOrDefault(reaperData.MAX_LVL, 0) : 1, 0, maxLevelAndAffinity ? valueOrDefault(reaperData.MAX_LVL, 0) : 1, 0, 0, 1000, maxLevelAndAffinity ? MAX_REAPER_AFFINITY_BONUS : 0));
+            .map(reaperData => this.slormancerReaperService.getReaper(
+                reaperData,
+                heroClass,
+                primordial,
+                maxLevelAndAffinity ? valueOrDefault(reaperData.MAX_LVL, 0) : 1,
+                0,
+                maxLevelAndAffinity ? valueOrDefault(reaperData.MAX_LVL, 0) : 1,
+                0,
+                0,
+                1000,
+                maxLevelAndAffinity ? (this.slormancerReaperService.useDifferentAffinityForEffects({ id: reaperData.REF, primordial: primordial } as any as Reaper) ? MAX_EFFECT_AFFINITY_BASE : MAX_REAPER_AFFINITY_BASE) : 0,
+                maxLevelAndAffinity ? MAX_REAPER_AFFINITY_BONUS : 0));
             
         this.filterReaperList();
     }